// src/Types/reducer-types.ts
import { CartItem, ShippingInfo, User } from "./Types";

export interface UserReducerInitialState {
  user: User;
  loading: boolean;
}

export interface CartReducerInitialState {
  loading: boolean;
  cartItems: CartItem[];
  subtotal: number;
  tax: number;
  shippingCharges: number;
  discount: number;
  total: number;
  shippingInfo: ShippingInfo;
}